backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
redis==5.0.1
openai==1.12.0
httpx==0.26.0
pydantic>=2.9.0
python-dotenv==1.0.0
numpy==1.26.3
feedparser==6.0.11
resend==2.1.0

